adding 'when' to prompts to determine when next prompt is applied. Allows next prompt without needing if/else or switch case
